+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Release notes
 All notable changes to this project will be documented in this file.
 
-## 0.8.0 WIP
+## 0.8.0
 - FIX: When splinterlands is in maintenance mode skip update, causing application crashes
 - Update readme for executable + change linux to ubuntu in github workflow
 - Change server mode wait time. After processing daily and season reschedule next pull cycle to 30 minutes.